fix update crash dagger
