+++ app/build.gradle
@@ -1,22 +1,19 @@
 apply plugin: 'com.android.application'
-
 apply plugin: 'kotlin-android'
-
 apply plugin: 'kotlin-android-extensions'
-
 apply plugin: 'com.apollographql.apollo'
-
 apply plugin: 'kotlin-kapt'
-
 apply plugin: 'org.jetbrains.kotlin.android.extensions'
 
 android {
-    compileSdkVersion 29
-    buildToolsVersion "29.0.3"
+
+    compileSdkVersion rootProject.compileSdkVersion
+    buildToolsVersion rootProject.buildToolsVersion
+
     defaultConfig {
         applicationId "com.example.demomvvmgraphql"
-        minSdkVersion 16
-        targetSdkVersion 29
+        minSdkVersion rootProject.minSdkVersion
+        targetSdkVersion rootProject.targetSdkVersion
         versionCode 1
         versionName "1.0"
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
@@ -34,9 +31,11 @@ android {
             ext.enableCrashlytics = false
         }
     }
+
     androidExtensions {
         experimental = true
     }
+
     kotlinOptions {
         jvmTarget = '1.8'
     }
@@ -57,10 +56,12 @@ android {
     dataBinding {
         enabled = true
     }
+
     compileOptions {
         sourceCompatibility JavaVersion.VERSION_1_8
         targetCompatibility JavaVersion.VERSION_1_8
     }
+
     flavorDimensions "flavor"
     productFlavors {
         chplay {
@@ -84,71 +85,41 @@ android {
 }
 
 dependencies {
-    implementation 'com.android.support:multidex:1.0.3'
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-    implementation 'androidx.appcompat:appcompat:1.1.0'
-    implementation 'androidx.core:core-ktx:1.2.0'
+    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
-    testImplementation 'junit:junit:4.13'
-    androidTestImplementation 'androidx.test.ext:junit:1.1.1'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.2.0'
-
-    implementation "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlin_version"
-
-    //Structure
-    implementation 'android.arch.lifecycle:extensions:1.1.1'
-    implementation 'com.android.support:multidex:1.0.3'
 
     //Design
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
-    implementation "com.airbnb.android:lottie:$lottieVersion"
-    implementation('com.github.thekhaeng:pushdown-anim-click:1.1.1') {
-        exclude group: 'com.android.support'
-    }
+    implementation 'com.airbnb.android:lottie:3.3.1'
+    implementation('com.github.thekhaeng:pushdown-anim-click:1.1.1')
 
     //Glide
     implementation 'com.github.bumptech.glide:glide:4.11.0'
-    kapt 'com.github.bumptech.glide:compiler:4.11.0'
 
     //Apollo
-    implementation "com.apollographql.apollo:apollo-runtime:$apolloVersion"
-    implementation "com.apollographql.apollo:apollo-rx2-support:$apolloVersion"
-    implementation "com.apollographql.apollo:apollo-android-support:$apolloVersion"
+    implementation 'com.apollographql.apollo:apollo-runtime:1.3.0'
+    implementation 'com.apollographql.apollo:apollo-rx2-support:1.3.0'
 
     //OkHttp
     implementation 'com.squareup.okhttp3:logging-interceptor:4.4.0'
 
     // Dagger 2
-    implementation "com.google.dagger:dagger:$dagger2_version"
-    implementation "com.google.dagger:dagger-android:$dagger2_version"
-    implementation "com.google.dagger:dagger-android-support:$dagger2_version"
-    kapt "com.google.dagger:dagger-compiler:$dagger2_version"
-    kapt "com.google.dagger:dagger-android-processor:$dagger2_version"
-    compileOnly "org.glassfish:javax.annotation:3.1.1"
-    implementation 'com.airbnb.android:lottie:3.3.1'
+    implementation 'com.google.dagger:dagger:2.26'
+    implementation 'com.google.dagger:dagger-android:2.26'
+    kapt 'com.google.dagger:dagger-compiler:2.26'
+    kapt 'com.google.dagger:dagger-android-processor:2.26'
+    implementation 'io.reactivex.rxjava2:rxkotlin:2.4.0'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+    compileOnly 'org.glassfish:javax.annotation:3.1.1'
 
-    //Rx
-    implementation "io.reactivex.rxjava2:rxjava:2.2.17"
-    implementation "io.reactivex.rxjava2:rxandroid:2.1.1"
-    implementation "io.reactivex.rxjava2:rxkotlin:2.4.0"
-
-    //Testing
-    implementation "junit:junit:$jUnitVersion"
-    implementation "com.nhaarman:mockito-kotlin:$mockitoKotlinVersion"
-    implementation "org.assertj:assertj-core:$assertJVersion"
-    implementation "com.google.code.gson:gson:2.8.6"
+    implementation 'com.google.code.gson:gson:2.8.6'
 
     implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.1.0'
 
-    implementation "androidx.recyclerview:recyclerview:1.1.0"
-    // For control over item selection of both touch and mouse driven selection
-    implementation "androidx.recyclerview:recyclerview-selection:1.1.0-rc01"
-
-    def lifecycle_version = "2.2.0"
 // ViewModel and LiveData
-    implementation "androidx.lifecycle:lifecycle-extensions:$lifecycle_version"
+    implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
 
-    implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
 
 }
