+++ app/src/main/java/com/example/demomvvmgraphql/base/Resource.kt
@@ -1,7 +1,11 @@
-package com.example.demomvvmgraphql.view.base
+package com.example.demomvvmgraphql.base
 
 sealed class Resource<out T> {
+
     class Loading<out T> : Resource<T>()
+
     data class Success<out T>(val data: T?) : Resource<T>()
+
     data class Failure<out T>(val cause: String?) : Resource<T>()
+
 }
