+++ app/src/main/java/com/example/demomvvmgraphql/binding/BindingAdapter.kt
@@ -1,14 +1,14 @@
-package com.tayfuncesur.pokehub.binding
+package com.example.demomvvmgraphql.binding
 
 import android.view.View
 import androidx.databinding.BindingAdapter
 
 object BindingAdapter {
+
     @JvmStatic
     @BindingAdapter("visibleGone")
     fun showHide(view: View, show: Boolean) {
         view.visibility = if (show) View.VISIBLE else View.GONE
     }
-
 
 }
