+++ app/src/main/java/com/example/demomvvmgraphql/module/AppComponent.kt
@@ -1,7 +1,7 @@
-package com.example.demomvvmgraphql.view.modul
+package com.example.demomvvmgraphql.module
 
 import android.app.Application
-import com.example.demomvvmgraphql.view.App
+import com.example.demomvvmgraphql.App
 import dagger.BindsInstance
 import dagger.Component
 import dagger.android.AndroidInjectionModule
