+++ app/src/main/java/com/example/demomvvmgraphql/module/AppInjector.kt
@@ -1,8 +1,9 @@
-package com.example.demomvvmgraphql.view.modul
+package com.example.demomvvmgraphql.module
 
-import com.example.demomvvmgraphql.view.App
+import com.example.demomvvmgraphql.App
 
 object AppInjector {
+
     fun init(app: App) {
         DaggerAppComponent.builder().application(app)
             .build().inject(app)
