+++ app/src/main/java/com/example/demomvvmgraphql/repository/PokemonDetailRepository.kt
@@ -1,24 +1,27 @@
+package com.example.demomvvmgraphql.repository
+
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
-import com.example.demomvvmgraphql.PokemonRepositoryQuery
-import com.example.demomvvmgraphql.view.base.Resource
-import com.example.demomvvmgraphql.view.repository.Repository
-import com.tayfuncesur.pokehub.network.ApiService
+import com.example.demomvvmgraphql.PokemonDetailRepositoryQuery
+import com.example.demomvvmgraphql.base.Resource
+import com.example.demomvvmgraphql.network.ApiService
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
 import io.reactivex.schedulers.Schedulers
 import javax.inject.Inject
+import javax.inject.Singleton
 
-class PokemonRepository @Inject constructor(private var apiService: ApiService) : Repository {
+@Singleton
+class PokemonDetailRepository @Inject constructor(private var apiService: ApiService) : Repository {
 
     private val compositeDisposable = CompositeDisposable()
 
-    private val result = MutableLiveData<Resource<List<PokemonRepositoryQuery.Pokemon>>>()
+    private val result = MutableLiveData<Resource<PokemonDetailRepositoryQuery.Pokemon>>()
 
-    fun getPokemons(count: Int): LiveData<Resource<List<PokemonRepositoryQuery.Pokemon>>> {
+    fun getPokemonDetail(id: String): LiveData<Resource<PokemonDetailRepositoryQuery.Pokemon>> {
         addDisposable(
-            apiService.getPokemons(count)
+            apiService.getPokemonDetail(id)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .doOnSubscribe { whenStart() }
@@ -35,8 +38,8 @@ class PokemonRepository @Inject constructor(private var apiService: ApiService) 
         result.value = Resource.Loading()
     }
 
-    fun whenSuccess(pokemonList: List<PokemonRepositoryQuery.Pokemon>) {
-        result.value = Resource.Success(pokemonList)
+    fun whenSuccess(pokemon: PokemonDetailRepositoryQuery.Pokemon) {
+        result.value = Resource.Success(pokemon)
     }
 
     fun whenError(cause: String) {
