+++ app/src/main/java/com/example/demomvvmgraphql/util/Const.kt
@@ -1,4 +1,4 @@
-package com.example.demomvvmgraphql.view.util
+package com.example.demomvvmgraphql.util
 
 var BASE_URL = "https://graphql-pokemon.now.sh/"
 const val SPLASH_SCREEN_DELAY = 1500L
