+++ app/src/main/java/com/example/demomvvmgraphql/viewmodel/ViewModelFactory.kt
@@ -1,4 +1,4 @@
-package com.example.demomvvmgraphql.view.viewmodel
+package com.example.demomvvmgraphql.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
@@ -11,10 +11,8 @@ import kotlin.reflect.KClass
 @Suppress("UNCHECKED_CAST")
 @Singleton
 class ViewModelFactory @Inject
-constructor(
-    private val creators: Map<Class<out ViewModel>,
-            @JvmSuppressWildcards Provider<ViewModel>>
-) : ViewModelProvider.Factory {
+constructor(private val creators: Map<Class<out ViewModel>, @JvmSuppressWildcards Provider<ViewModel>>) :
+    ViewModelProvider.Factory {
 
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
         var creator: Provider<out ViewModel>? = creators[modelClass]
