+++ app/src/main/res/layout/activity_main.xml
@@ -7,13 +7,13 @@
 
         <import type="androidx.lifecycle.LiveData" />
 
-        <import type="com.example.demomvvmgraphql.view.base.Resource.Loading" />
+        <import type="com.example.demomvvmgraphql.base.Resource.Loading" />
 
         <import type="com.example.demomvvmgraphql.PokemonRepositoryQuery.Pokemon" />
 
-        <import type="com.example.demomvvmgraphql.view.base.Resource.Failure" />
+        <import type="com.example.demomvvmgraphql.base.Resource.Failure" />
 
-        <import type="com.example.demomvvmgraphql.view.base.Resource" />
+        <import type="com.example.demomvvmgraphql.base.Resource" />
 
         <import type="java.util.List" />
