+++ app/src/main/res/layout/activity_spash_screen.xml
@@ -3,6 +3,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.activity.SpashActivity">
+    tools:context=".activity.SpashActivity">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
