+++ build.gradle
@@ -2,14 +2,6 @@
 
 buildscript {
     ext.kotlin_version = '1.3.61'
-    ext.retrofit_version = '2.4.0'
-    ext.dagger2_version = '2.26'
-    ext.lottieVersion = '2.7.0'
-    ext.apolloVersion = '1.3.0'
-    ext.jUnitVersion = '4.12'
-    ext.mockitoKotlinVersion = '1.6.0'
-    ext.assertJVersion = '3.12.2'
-    ext.gsonVersion = '2.8.5'
 
     repositories {
         google()
@@ -17,13 +9,11 @@ buildscript {
         jcenter()
 
     }
+
     dependencies {
         classpath 'com.android.tools.build:gradle:3.5.3'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-        classpath "com.apollographql.apollo:apollo-gradle-plugin:$apolloVersion"
-
-        // NOTE: Do not place your application dependencies here; they belong
-        // in the individual module build.gradle files
+        classpath 'com.apollographql.apollo:apollo-gradle-plugin:1.3.0'
     }
 }
 
@@ -38,4 +28,15 @@ allprojects {
 
 task clean(type: Delete) {
     delete rootProject.buildDir
+}
+
+ext {
+    //dependencies
+    junitVersion = '4.12'
+    espressoVersion = '2.2.2'
+    minSdkVersion = 16
+    compileSdkVersion = 29
+    targetSdkVersion = 29
+    buildToolsVersion = '29.0.3'
+    supportLibraryVersion = '29.0.0'
 }
