+++ gradle.properties
@@ -14,8 +14,6 @@ org.gradle.jvmargs=-Xmx1536m
 # AndroidX package structure to make it clearer which packages are bundled with the
 # Android operating system, and which are packaged with your app's APK
 # https://developer.android.com/topic/libraries/support-library/androidx-rn
-android.useAndroidX=true
 # Automatically convert third-party libraries to use AndroidX
-android=enableJetifier=true
-# Kotlin code style for this project: "official" or "obsolete":
-kotlin.code.style=official
+android.useAndroidX=true
+android.enableJetifier=true
